create component for modal content
