+++ src/components/Modals/Projects/ModalLoyaltyApplication.css
@@ -1,0 +1,3 @@
+.modal-loyalty.box:nth-child(2) {
+  width: 50vw;
+}
